+++ tests/godot_plugin_harness_fixture/tests/tool_localization_inventory_contract_test.gd
@@ -10,6 +10,8 @@ const ToolPresentationServiceScript = preload("res://addons/godot_dotnet_mcp/plu
 const REMOVED_PUBLIC_TOOL_LOCALIZATION_KEYS: Array[String] = [
 	"tool_system_help_name",
 	"tool_system_help_desc",
+	"tool_system_tool_catalog_name",
+	"tool_system_tool_catalog_desc",
 	"tool_system_tool_activity_name",
 	"tool_system_tool_activity_desc",
 	"tool_system_scene_validate_name",